Update Dice class to use uint64_t for seed and distribution: Modify constructor and seed method to accept uint64_t, ensuring consistency with game configuration. Now games are different in the same simulation but with the same seed every game is the same between simulations

+++ src/extra/Dice.hpp
@@ -10,9 +10,9 @@ class Dice {
         std::uniform_int_distribution<int> _dieDistribution{1, 6};
     public:
         Dice() : _generator(std::random_device{}()) {}
-        Dice(uint32_t seed) : _generator(seed) {}
+        Dice(uint64_t seed) : _generator(seed) {}
 
-        void seed(uint32_t seed) {
+        void seed(uint64_t seed) {
             _generator.seed(seed);
         }
